Add channel=release|next
* Also fix regex for opam option setenv
* Replace end_pkg_vers and post_pkg_vers with end_pkgs and post_pkg to
  reflect that the versions may be different after pinning

+++ ci/prepare-release.sh
@@ -1,18 +1,22 @@
 #!/bin/sh
 set -euf
 
-# Dump all ABIs into .ci/stage-release
-install -d .ci
-ARCHIVE="$(pwd)/.ci/stage-release"
-rm -rf "$ARCHIVE"
-install -d "$ARCHIVE"
-find dist -mindepth 1 -maxdepth 1 -type f -name "*.tar.gz" | while read -r tarball; do
-    dkml_host_abi=$(basename "${tarball%.tar.gz}")
-    install -d "$ARCHIVE/$dkml_host_abi"
-    tar xCvfz "$ARCHIVE/$dkml_host_abi" "$tarball"
-done
-
-# Tar ball
 rm -rf "_release"
 install -d "_release"
-tar cvCfz "$ARCHIVE" "_release/staging.tar.gz" .
+
+for CHANNEL in release next; do
+    if [ -d "dist/$CHANNEL" ]; then
+        # Dump all ABIs into .ci/<channel>stage-release
+        ARCHIVE="$(pwd)/.ci/$CHANNEL/stage-release"
+        rm -rf "$ARCHIVE"
+        install -d "$ARCHIVE"
+        find "dist/$CHANNEL" -mindepth 1 -maxdepth 1 -type f -name "*.tar.gz" | while read -r tarball; do
+            dkml_host_abi=$(basename "${tarball%.tar.gz}")
+            install -d "$ARCHIVE/$dkml_host_abi"
+            tar xCvfz "$ARCHIVE/$dkml_host_abi" "$tarball"
+        done
+
+        # Tar ball
+        tar cvCfz "$ARCHIVE" "_release/$CHANNEL.tar.gz" .
+    fi
+done
